storyboard/requirements.txt
appdirs==1.4.4
audioread==3.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
decorator==5.1.1
einops==0.6.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.2
fire==0.5.0
ftfy==6.1.1
future==0.18.3
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
idna==3.4
imageio==2.31.1
importlib-metadata==4.8.1
importlib-resources==6.0.1
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.0.post2
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numba==0.57.1
numpy==1.24.4
openai-whisper==20230314
packaging==23.1
Pillow==10.0.0
pooch==1.6.0
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pytorch-ranger==0.1.1
regex==2023.8.8
requests==2.31.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
siren-pytorch==0.1.7
six==1.16.0
soundfile==0.12.1
soxr==0.3.6
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.3.1
torch==2.0.1
torch-optimizer==0.3.0
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6
zipp==3.16.2
python-dotenv==1.0.0